some changes about configurations
